Create BOM/STL readme, add folder download link
Adds a readme to the "Assemblies BOM and STL" folder; Adds a small section that includes a link to the download-directory.github.io tool, which allows downloading the entire directory without downloading the repository or requiring a downloadable ZIP file. This tool zips the folder up automatically, and is client-side JavaScript powered.